requirements.txt
alembic==1.11.2
beautifulsoup4==4.9.3
datasets==2.14.4
fire==0.5.0
Flask==2.3.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.5
googlesearch_python==1.2.3
lit_nlp==0.5.0
mpi4py==3.1.4
nltk==3.8.1
numpy==1.24.4
omegaconf==2.3.0
openai==0.27.0
openpyxl==3.1.2
pandas==2.0.3
pydantic==1.10.12
pytest==7.4.0
pytorch_lightning==2.0.7
PyYAML==5.4.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.11.2
spacy==3.6.1
SQLAlchemy==2.0.19
torch==2.0.1
tqdm==4.65.2
transformers==4.29.2
